interactive solution to the NucleoPebble game
To find out a winning strategy, first we need to play some matches. We deal with winning positions and loosing positions, that characterize the practical strategy of the game of Nim. Indeed, the purposes of Nucleopebble are not the same as Nim, thus, we need something more, to check if our sequence is a uni-nucleotide one.